Remove all ICU integration for sqlite on host
Maintaining external/icu for non-device usecases is
becoming an increasing burden and there are no tests
around sqlite / libsqlite on host that indicate what
support is needed.

There doesn't appear to be any ICU initialization code
in place on host suggesting ICU usage is limited to
functions that don't require data.

This change removes ICU integration with sqlite on host
entirely to try to flush out affected users. If there
are no complaints that will be one less dependency to
worry about.

Bug: 130787951
Test: build / treehugger
Test: Ran sqlite3 on host
Change-Id: Icd1eb97281d941a3f55c4289b6d65fe5921ed542