sqlite: Upgrade to SQLite 3.32.2
This change is a cherry-pick of aosp/1337179 which is being merged here
because of merge conflicts due to changes [1] and [2] which were made
in internal master and rvc-dev but not in AOSP.

Downloaded from
https://www.sqlite.org/2020/sqlite-autoconf-3320200.tar.gz

$ sha1sum sqlite-autoconf-3320200.tar.gz
429e3f2d0b16a95ad1025a97b2a328d0b4037575  sqlite-autoconf-3320200.tar.gz

dist/orig contains the stock sqlite3 code, as packaged in the tar.gz file above.

dist contains a copy of dist/orig, but with the Android.patch file applied.

More details about the release are available here:
https://www.sqlite.org/releaselog/3_32_2.html

[1] ag/49895f5e2490f20a532b4492648006b0fd29ff7e
[2] ag/e7d688ce246a1a7fa757dcd3716ff5f1a5d921b8

Test: code compiles and device boots with no problems.
Test: select sqlite_version() - returns 3.32.2
Test: atest cts/SQLiteDatabaseTest all passed.
Test: atest cts/SQLiteSecurityTest
Bug: 159105124
Change-Id: I5c5fa2fd90292483bf20d8ae58adad41c48aa4e6