Add libsqlite_static_noicu
The static library is variant of libsqlite built without the ICU
extension. The library is supposed to be used in environments where the
ICU extension is not needed and the dependency to ICU (which is as large
as 60+ MB) is not desirable, like microdroid.

Bug: 199674764
Test: atest MicrodroidHostTestCases
Merged-In: Id8bfe752f70d990aa6cd5a4c121a2cc0541063b5
Change-Id: Id8bfe752f70d990aa6cd5a4c121a2cc0541063b5