Make Windows variant of libsqlite consistent with Linux+Mac
Previously, the Linux and Mac variants of libsqlite were compiled with
-DSQLITE_ENABLE_ICU. The Windows variant did not include those flags.
Update the libsqlite Windows variant to include the flags.

Windows originally did not include icu because external/icu was not
enabled for host Windows, but it now builds for host Windows.

Host libsqlite is used in tools like Robolectric and LayoutLib.

Note that there is still a libsqlite_static_noicu that contains a
variant of sqlite that is not build with ICU extensions.

Test: m sqlite3 py2-launcher py3-launcher py3-launcher-autorun
Test: m traceconv trace_processor_shell
Bug: 308173442
Change-Id: Ie5f700bb1d8e960ea13cbe588d0fdd747264591f