Consume build flag in sqlite Android.bp
The two Android.bp files for sqlite now use the build flag
RELEASE_PACKAGE_LIBSQLITE3.  The value of the build flag must be a
seven-digit sqlite release number.  For example, release 3.44.2 is
coded as 3440200.  This is the same number that sqlite.org uses for
the release URLs.

The legacy source files are removed.

Bug: 323176655

Change-Id: Ib75fcbd7f9eac64335be48fd6a49cf67765b8c1b